==> src/operators/R/3_validation.R <==
# Author: Brayden Tang
# Date: May 20th, 2020

"This script validates both the best simple GLM and the best Bayesian model
against the held out test set. This script assumes it will be run from the 
root of the repository. Furthermore, this script also generates Bayesian
diagnostic plots and summaries to ensure legitimacy of MCMC samples, if need be.

Usage: 3_validation.R --path_to_glm=<path_to_glm> --path_to_bayesian=<path_to_bayesian> <test_data_path> <results_out> <path_out_diag>

Options:
--path_to_glm=<path_to_glm>             A file path to the best basic GLM model.
--path_to_bayesian=<path_to_bayesian>   A file path to the best Bayesian model.
<test_data_path>                        A file path to the test data set.
<results_out>                           A file path specifying where to output validation results. This should NOT be a specific file (like .rds or .csv), just a directory.
<path_out_diag>                         A file path specifying where to output MCMC diagnostic plots and tables. This should NOT be a specific file (like .rds or .csv), just a directory.
" -> doc

library(tidyverse)
library(parallel)
library(bayesplot)
library(brms)
library(docopt)

opt <- docopt(doc)

#' This function checks a brms Bayesian model for convergence, mixing,
#' and any other violations that would yield to questioning of the samples
#' generated by the MCMC process.
#' 
#' @param model A brms fitted model object. 
#' @param path_out_diag A file path specifying where to output diagnostics.
#'
#' @return None
#' @export
#'
#' @examples
#' check_samples(
#' model = my_brms_Fitted_model,
#' path_out_diag = "results/operators/bayes-diagnostics"
#' )
check_samples <- function(model, path_out_diag) {
  
  # Input validation - check that model is brmsfit and that output path is not 
  # an actual file.
  if (!is.brmsfit(model)) {
    stop("The model object must be of brmsfit.")
  } else if (str_detect(path_out_diag, "\\.xlsx$|\\.csv$|\\.rds$|//.txt$")) {
    stop("The output path for model diagnostics must be a general path, not a 
         path to a specific file. Remove the file extension.")
  }
  
  trace_plot_global <- mcmc_trace(
    model,
    pars = c(parnames(model)[1:4])
    )
  
  acf_plot_global <- mcmc_acf(
    model,
    pars = c(parnames(model)[1:4])
  )
  
  if (!dir.exists(path_out_diag)) {
  	dir.create(path_out_diag)
  }
  
  ggsave(trace_plot_global, filename = paste0(path_out_diag, "/trace_plot_globals.png"))
  ggsave(acf_plot_global, filename = paste0(path_out_diag, "/acf_plot_globals.png"))
  
  saveRDS(summary(model)$fixed, paste0(path_out_diag, "/summary_stats_bayes.rds"))
  
}

#' Validates the best GLM and Bayesian models against a held out test set,
#' and then prints the results. Also, outputs diagnostic plots of the MCMC chain.
#'
#' @param path_to_glm A file path to the best GLM model.
#' @param path_to_bayesian A file path to the best Bayesian model.
#' @param test_data_path A file path to the test data set.
#' @param results_out A file path specifying where to output the test set scores. 
#' Should only be a directory, not a specific file name.
#' @param path_out_diag A file path specifying where to output MCMC diagnostic plots. 
#' Should only be a directory, not a specific file name.
#'
#' @return None
#' @export
#'
#' @examples
#' main(
#' path_to_glm = "results/operators/models/basic-glm.rds",
#' path_to_bayesian = "results/operators/models/best-bayes-model.rds",
#' test_data_path = "data/operators/test.csv",
#' results_out = "results/operators/report-tables",
#' path_out_diag = "results/operators/bayes-diagnostics"
#' )
main <- function(path_to_glm, path_to_bayesian, test_data_path, results_out, path_out_diag) {

  # Input validation checks. Checks for correctly specified file paths.
  if (!str_detect(path_to_glm, ".rds")) {
    stop("File path to GLM must be a specific file with extension .rds")
  } else if (!str_detect(path_to_bayesian, ".rds")) {
    stop("File path to Bayesian model must be a specific file with extension .rds")
  } else if (!str_detect(test_data_path, ".csv")) {
    stop("File path to test set must be a specific file with extension .csv")
  } else if (str_detect(results_out, "\\.txt$|\\.csv$|\\.xlsx$|//.rds$")) {
  	stop("results_out should just be a directory, not a specific file. Remove the file extension.")
  } else if (str_detect(path_out_diag, "\\.txt$|\\.csv$|\\.xlsx$|//.rds$")) {
    stop("path_out_diag should just be a directory, not a specific file. Remove the file extension.")
  } else if (endsWith(results_out, "/")) {
  	stop("File path for results_out should not end with /")
  } else if (endsWith(path_out_diag, "/")) {
  	stop("File path for path_out_diag should not end with /")
  } 
  
  final_bayes_model <- readRDS(path_to_bayesian)
  final_glm <- readRDS(path_to_glm)
  
  # Run the model diagnostics first on the Bayesian model.
  
  check_samples(model = final_bayes_model, path_out_diag = path_out_diag)
  
  # Now, validate the model against the test set.
  
  test_data <- read_csv(test_data_path) %>%
    mutate(hours_worked_div_1957 = total_hours_last3yr / 1957)
  
  test_results_glm <- predict(final_glm, newdata = test_data, type = "response")
  set.seed(200350623)
  test_results_bayes <- predict(final_bayes_model, newdata = test_data)
  
  # Store the results in a table format.
  
  results <- tibble(
    model = c("Best Simple GLM", "Best Bayesian"),
    MAE = c(mean(abs(test_results_glm - test_data$number_incidents)), 
            mean(abs(test_results_bayes[, 1] - test_data$number_incidents)))
    )
  
  if (!dir.exists(results_out)) {
  	dir.create(results_out)
  }
  
  # Save the table of test set scores.
  
  saveRDS(results, paste0(results_out, "/validation-results.rds"))

}

main(
  path_to_glm = opt$path_to_glm,
  path_to_bayesian = opt$path_to_bayesian,
  test_data_path = opt$test_data_path,
  results_out = opt$results_out,
  path_out_diag = opt$path_out_diag
)
